0x1A. C - Hash tables

+++ 0x1A-hash_tables/2-key_index.c
@@ -11,7 +11,7 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int index = hash_djb2(key) % size;
+	unsigned long int index = hash_djb2(key) % size;
 
 	return (index);
 }
